OK-579: Refaktoroitu "sijoitteluajon tulokset valintatiedoilla"-datan konvertointi ulos api-funktiosta, jotta ei tarvitse noutaa hakemuksia aina samalla kun halutaan hakea uudelleen sijoitteluajon tulokset.
