+++ src/app/haku/[oid]/hakukohde/[hakukohde]/sijoittelun-tulokset/lib/refetch-sijoittelun-tulokset.ts
@@ -1,4 +1,4 @@
-import { tryToGetLatestSijoitteluajonTuloksetWithValintaEsitysQueryOptions } from '@/lib/valinta-tulos-service/valinta-tulos-service';
+import { getLatestSijoitteluajonTuloksetWithValintaEsitysQueryOptions } from '@/lib/valinta-tulos-service/valinta-tulos-service';
 import { QueryClient } from '@tanstack/react-query';
 
 export const refetchSijoittelunTulokset = (
@@ -6,11 +6,10 @@ export const refetchSijoittelunTulokset = (
   hakukohdeOid: string,
   queryClient: QueryClient,
 ) => {
-  const options =
-    tryToGetLatestSijoitteluajonTuloksetWithValintaEsitysQueryOptions({
-      hakuOid,
-      hakukohdeOid,
-    });
+  const options = getLatestSijoitteluajonTuloksetWithValintaEsitysQueryOptions({
+    hakuOid,
+    hakukohdeOid,
+  });
   queryClient.invalidateQueries(options);
   queryClient.resetQueries(options);
 };
