+++ src/lib/common.ts
@@ -104,7 +104,7 @@ export const nullWhen404 = async <T>(
   } catch (e) {
     if (e instanceof FetchError && e?.response?.status === 404) {
       console.error('FetchError with 404', e);
-      return null;
+      return Promise.resolve(null);
     }
     throw e;
   }
